pyproject.toml
[project]
name = "mcp-server-starrocks"
version = "0.3.0"
description = "official MCP server for StarRocks"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "fastmcp>=2.14.0",
    "mysql-connector-python>=9.2.0",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "kaleido==0.2.1",
    "adbc-driver-manager>=0.8.0",
    "adbc-driver-flightsql>=0.8.0",
    "pyarrow>=14.0.0",
    "PyJWT[crypto]>=2.8.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[[project.authors]]
name = "changbinglin"
email = "changbinglin@starrocks.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-server-starrocks = "mcp_server_starrocks:main"

[project.urls]
Home = "https://github.com/starrocks/mcp-server-starrocks"
